// app-ui/src/app/backend/model/investigationCreateRequest.ts
/**
 * TIBCO Discover public API
 * TIBCO Discover public API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InvestigationCreateRequest { 
    type?: string;
    ids?: string;
    details?: string;
    summary?: string;
    analysisId?: string;
    analysisName?: string;
    templateId?: string;
    templateName?: string;
}